refactor(gui-v2): use page for shared view form
