+++ packages/nc-gui-v2/pages/[projectType]/form/[viewId].vue
@@ -1,9 +1,19 @@
 <script setup lang="ts">
 import type { Ref } from 'vue'
-import type { TableType } from 'nocodb-sdk/build/main'
-import { useProvideSharedFormStore } from '~/composables/useSharedFormViewStore'
-import { IsFormInj, IsPublicInj, MetaInj, ReloadViewDataHookInj } from '~/context'
-import { createEventHook, definePageMeta, provide, ref, useProvideSmartsheetStore, useRoute } from '#imports'
+import type { TableType } from 'nocodb-sdk'
+import {
+  IsFormInj,
+  IsPublicInj,
+  MetaInj,
+  ReloadViewDataHookInj,
+  createEventHook,
+  definePageMeta,
+  provide,
+  ref,
+  useProvideSharedFormStore,
+  useProvideSmartsheetStore,
+  useRoute,
+} from '#imports'
 
 definePageMeta({
   public: true,
@@ -11,13 +21,12 @@ definePageMeta({
 
 const route = useRoute()
 
-const reloadEventHook = createEventHook<void>()
-
 const { loadSharedView, sharedView, meta, notFound } = useProvideSharedFormStore(route.params.viewId as string)
 
 await loadSharedView()
+
 if (!notFound.value) {
-  provide(ReloadViewDataHookInj, reloadEventHook)
+  provide(ReloadViewDataHookInj, createEventHook())
   provide(MetaInj, meta)
   provide(IsPublicInj, ref(true))
   provide(IsFormInj, ref(true))
@@ -27,5 +36,7 @@ if (!notFound.value) {
 </script>
 
 <template>
-  <SharedViewForm />
+  <NuxtLayout>
+    <NuxtPage />
+  </NuxtLayout>
 </template>
